+++ src/main/java/com/nisovin/shopkeepers/lang/Messages.java
@@ -315,7 +315,7 @@ public class Messages extends Config {
 	public static Text commandDescriptionReload = Text.parse("Reloads this plugin.");
 	public static Text commandDescriptionDebug = Text.parse("Toggles debug mode on and off.");
 	public static Text commandDescriptionList = Text.parse("Lists all shops for the specified player, or all admin shops.");
-	public static Text commandDescriptionRemove = Text.parse("Removes all shops for the specified player, all players, or all admin shops.");
+	public static Text commandDescriptionRemoveAll = Text.parse("Removes all shops of a specific player, all players, or all admin shops.");
 	public static Text commandDescriptionGive = Text.parse("Gives shop creation item(s) to the specified player.");
 	public static Text commandDescriptionGiveCurrency = Text.parse("Gives currency item(s) to the specified player.");
 	public static Text commandDescriptionConvertItems = Text.parse("Converts the held (or all) items to conform to Spigot's data format.");